dedupe code snippet
